+++ param_global/hooks.py
@@ -42,6 +42,8 @@ app_include_js = [
 	"pg_grille_entree.bundle.js",
 	"pg_tri_dropdown.bundle.js",
 	"pg_environnement.bundle.js",
+	"pg_init_doctype.bundle.js",
+	"pg_controle_date.bundle.js",
 ]
 
 # include js, css files in header of web template